Refactor Shopily workspace presenter
